requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.10.0
iniconfig==1.1.1
Markdown==3.3.6
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.2
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
zipp==3.6.0
